requirements.txt
Flask
Flask-SQLAlchemy
Authlib
confluent-kafka
psycopg2-binary
FastAPI
uvicorn
httpx
pyjwt
jsonschema
